# cmd/api/docs/swagger.yaml
definitions:
...
  domain.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      permission:
        type: boolean
      premium:
        type: boolean
    required:
    - email
    - name
    - password
    type: object
  handler.Otp:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      err:
        type: string
      hint:
        type: string
      status:
        type: string
    type: object
...
  models.SignupData:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
  models.UpdateData:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Userlogindata:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /login:
    get:
      description: Retrieve the HTML page for user login.
      produces:
      - application/json
      responses:
        "200":
          description: Got HTML page for user login
          schema:
            type: string
      summary: Get HTML page for user login
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Log in a user and return a token if successful.
      parameters:
      - description: User login data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Userlogindata'
      produces:
      - application/json
      responses:
        "200":
          description: 'succesfully logged in  + tokenstring '
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - users
  /otplogin:
    get:
      description: Retrieve the HTML page for OTP login.
      produces:
      - application/json
      responses:
        "200":
          description: Got HTML page for OTP login, enter email
          schema:
            type: string
      summary: Get HTML page for OTP login
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Verify the provided email and send an OTP.
      parameters:
      - description: User email to verify and send OTP
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified, OTP sent  + Email
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify email and send OTP
      tags:
      - users
  /signup:
    get:
      description: Retrieve the HTML page for user signup.
      produces:
      - text/html
      responses:
        "200":
          description: Successfully got the HTML page
          schema:
            type: string
      summary: Get HTML page for user signup
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Save a user by providing JSON payload
      operationId: SaveUser
      parameters:
      - description: User object to be saved
        in: body
        name: usersdata
        required: true
        schema:
          $ref: '#/definitions/models.SignupData'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: InternalServerError
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Save a user
      tags:
      - users
  /user/account/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account by providing the user ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User is deleted successfully || Redirected to login page
          schema:
            type: string
        "400":
          description: Bad Request, couldn't parse id
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Couldn't delete user
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a user account
      tags:
      - users
...
  /user/profile/{id}:
    get:
      consumes:
      - application/json
      description: Get user profile information by providing the user ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request, couldn't parse id
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Couldn't get user profile
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user profile
      tags:
      - users
  /user/update:
    put:
      consumes:
      - application/json
      description: Update a user's information
      parameters:
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a user
      tags:
      - users
...
  /verifyotp:
    post:
      consumes:
      - application/json
      description: Verify the provided OTP.
      parameters:
      - description: OTP data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.Otp'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified, redirected to home page
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized OTP
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify OTP
      tags:
      - users
swagger: "2.0"